0x13. C - More singly linked lists

+++ 0x13-more_singly_linked_lists/1-listint_len.c
@@ -7,7 +7,7 @@
  */
 size_t listint_len(const listint_t *h)
 {
-	size_t my_count;
+	unsigned int my_count;
 
 	if (h == NULL)
 	{
